agents/AwsKnowledgeAgent/app/AWSKnowledgeAgent/pyproject.toml
[project]
name = "AWSKnowledgeAgent"
version = "0.1.0"
description = "AgentCore Runtime Application using Strands SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "bedrock-agentcore >= 1.0.3",
  "mcp >= 1.19.0",
  "strands-agents >= 1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
